+++ NULL Engine/Resources/Assets/Scripts/Turret.h
@@ -50,7 +50,7 @@ public:
 	C_AudioSource* deathAudio = nullptr;
 
 	int minCredits = 0.f;
-	float maxCredits = 0.f;
+	int maxCredits = 0.f;
 
 private:
